[Refactor] : 그래프 컴포넌트 분리
