+++ src/components/vote/VoteGraph.tsx
@@ -1,0 +1,22 @@
+import voteGarphCalc from '../../utils/voteGraphCalc';
+
+function VoteGraph() {
+  const { graph, graphTotal } = voteGarphCalc();
+
+  return (
+    <section className="vote-result">
+      <ul className="vote-result__graph">
+        {graph.map((item) => (
+          <li
+            style={{ width: `${(item / graphTotal) * 100}%` }}
+            className="vote-result__graph-item"
+          >
+            <div>{item}</div>
+          </li>
+        ))}
+      </ul>
+    </section>
+  );
+}
+
+export default VoteGraph;
