+++ src/utils/voteGraphCalc.tsx
@@ -1,0 +1,6 @@
+export default function voteGarphCalc() {
+  const graph: number[] = [1, 2, 3, 4];
+  const graphTotal: number = graph.reduce((arr, cur) => arr + cur, 0);
+
+  return { graph, graphTotal };
+}
